# docs/swagger.yaml
basePath: /
definitions:
...
  entities.Teacher:
    properties:
      accountId:
        type: integer
      fio:
        type: string
      id:
        type: integer
      phoneNumber:
        type: string
    type: object
...
  requests.CreateTeacherRequest:
    properties:
      account_id:
        type: integer
      fio:
        type: string
      phone_number:
        type: string
    type: object
...
  requests.UpdateTeacherRequest:
    properties:
      account_id:
        type: integer
      fio:
        type: string
      id:
        type: integer
      phone_number:
        type: string
    type: object
...
  usecases.ReadTeacherResponseDto:
    properties:
      teacher:
        $ref: '#/definitions/entities.Teacher'
    type: object
host: localhost:8000
info:
  contact: {}
  description: Backend for KeenEye
  title: Backend for KeenEye
  version: 1.0.0
paths:
...
  /api/create-teacher:
    post:
      consumes:
      - application/json
      description: Create a new teacher (admin only)
      parameters:
      - description: Teacher info
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/requests.CreateTeacherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Teacher'
        "400":
          description: Invalid request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Create teacher
      tags:
      - teachers
...
  /api/delete-teacher:
    delete:
      description: Delete teacher by ID (admin only)
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid teacher ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Delete teacher
      tags:
      - teachers
...
  /api/read-all-teachers:
    get:
      description: Returns list of all teachers (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Teacher'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get all teachers
      tags:
      - teachers
...
  /api/read-teacher:
    get:
      description: Get teacher by ID (teacher sees self, admin sees all, students
        forbidden)
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecases.ReadTeacherResponseDto'
        "400":
          description: Invalid teacher ID
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Get teacher by ID
      tags:
      - teachers
...
  /api/update-teacher:
    put:
      consumes:
      - application/json
      description: Update teacher info (teacher updates self, admin updates any)
      parameters:
      - description: Updated teacher info
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateTeacherRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Teacher'
        "400":
          description: Invalid request
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: object
        "403":
          description: Forbidden
          schema:
            type: object
        "500":
          description: Internal server error
          schema:
            type: object
      security:
      - BasicAuth: []
      summary: Update teacher
      tags:
      - teachers
securityDefinitions:
  BasicAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"